test(paraswap): refactor and cleanup
