+++ test/paraswap-v5/paraswap-v5.ts
@@ -1,9 +1,9 @@
 import { TransactionRequest } from "@ethersproject/abstract-provider";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { expect } from "chai";
-import { BigNumber, BytesLike, Contract, ContractFactory } from "ethers";
+import { BigNumber, Contract, ContractFactory } from "ethers";
 import { ethers } from "hardhat";
-import { ContractMethod, ParaSwap } from "paraswap";
+import { getTransaction, signCalldata } from "./paraswap.service";
 
 const ETH_ADDRESS = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
 const DAI_ADDRESS = "0x6B175474E89094C44Da98b954EedeAC495271d0F";
@@ -13,51 +13,7 @@ const REGISTRY_ID = 0;
 const TIMELOCK = 0;
 
 const ethAmount = ethers.utils.parseEther("0.1").toString();
-const paraSwap = new ParaSwap();
-
-const getTransaction = async (sender: string) => {
-  const srcToken = ETH_ADDRESS;
-  const destToken = DAI_ADDRESS;
-  const srcAmount = ethAmount;
-  const referrer = "argent";
-
-  const priceRoute = await paraSwap.getRate(
-    srcToken, 
-    destToken, 
-    srcAmount,
-    undefined,
-    undefined,
-    { includeContractMethods: [ContractMethod.simpleSwap] }
-  );
-
-  if ("message" in priceRoute) {
-    throw new Error(priceRoute.message);
-  }
-
-  const transaction = await paraSwap.buildTx(
-    srcToken,
-    destToken,
-    srcAmount,
-    BigNumber.from(priceRoute.destAmount).mul("99").div("100").toString(),
-    priceRoute,
-    sender,
-    referrer,
-    undefined,
-    undefined,
-    undefined,
-    { ignoreChecks: true },
-  );
-
-  if ("message" in transaction) {
-    throw new Error(`Failed to build transaction: ${transaction.message}`);
-  }
-
-  transaction.value = BigNumber.from(transaction.value).toHexString();
-  // @ts-ignore
-  transaction.gasPrice = BigNumber.from(transaction.gasPrice).toHexString();
-
-  return transaction;
-};
+const daiAmount = ethers.utils.parseEther("100").toString();
 
 describe("Paraswap", function () {
   let DappRegistry: ContractFactory;
@@ -77,15 +33,6 @@ describe("Paraswap", function () {
     registry = await DappRegistry.deploy(TIMELOCK);
     [deployer, wallet, ags] = await ethers.getSigners();
   });
-
-  const signCalldata = async (data?: BytesLike) => {
-    if (!data) {
-      throw new Error("No calldata");
-    }
-    const messageHash = ethers.utils.keccak256(data);
-    const messageBytes = ethers.utils.arrayify(messageHash);
-    return await ags.signMessage(messageBytes);
-  };
 
   describe("Testing filter for Paraswap v5", function () {
     let augustus: any;
@@ -107,8 +54,9 @@ describe("Paraswap", function () {
 
       filter = await ParaswapFilter.deploy(ags.address);
       await registry.addDapp(REGISTRY_ID, forwarder.address, filter.address);
-      ethToDaiTransaction = await getTransaction(deployer.address);
+
       dai = await ethers.getContractAt("IERC20", DAI_ADDRESS);
+      ethToDaiTransaction = await getTransaction(ETH_ADDRESS, dai.address, ethAmount, deployer.address);
     });
 
     it("Should accept swapOnUniswap", async function () {
@@ -116,7 +64,7 @@ describe("Paraswap", function () {
       const amountOutMin = "1";
       const path = [ethers.constants.AddressZero];
       const swapData = augustus.interface.encodeFunctionData("swapOnUniswap", [amountIn, amountOutMin, path]);
-      const signature = await signCalldata(swapData);
+      const signature = await signCalldata(swapData, ags);
 
       const data = forwarder.interface.encodeFunctionData("delegateToParaswap", [signature, swapData]);
       const isAuthorised = await registry.isAuthorised(wallet.address, forwarder.address, forwarder.address, data);
@@ -131,7 +79,7 @@ describe("Paraswap", function () {
     });
 
     it("Should swap ETH for DAI on augustus using the forwader", async function () {
-      const signature = await signCalldata(ethToDaiTransaction.data);
+      const signature = await signCalldata(ethToDaiTransaction.data, ags);
       const balanceBefore = await dai.balanceOf(deployer.address);
       await forwarder.delegateToParaswap(signature, ethToDaiTransaction.data, { value: ethAmount });
       const balanceAfter: BigNumber = await dai.balanceOf(deployer.address);
